% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascollDB.R
\name{as.collDB}
\alias{as.collDB}
\title{function to convert a list into a collDB}
\usage{
as.collDB(document)
}
\arguments{
\item{document}{The document to be analysed}
}
\description{
function to convert a list into a collDB
}
