% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CollDB.R
\docType{class}
\name{collDB-class}
\alias{collDB-class}
\alias{collDB}
\title{a class containing a list of collocate information, derived from the 'save_collocates' function}
\description{
a class containing a list of collocate information, derived from the 'save_collocates' function
}
