% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collDB_test.R
\name{collDB_test}
\alias{collDB_test}
\title{Tests input document and returns in collDB if required}
\usage{
collDB_test(document, window, node, remove_stops = TRUE,
  remove_numerals = TRUE, remove_punct = TRUE)
}
\arguments{
\item{document}{The document to be analysed}

\item{window}{The size of the collocate window on each side of tne node}

\item{node}{A key word or phrase}

\item{remove_stops}{If TRUE, stopwords are removed (stopwords derived from quanteda package)}

\item{remove_numerals}{If TRUE, numerals are removed}

\item{remove_punct}{If TRUE, puntuation is removed}
}
\description{
Tests input document and returns in collDB if required
}
