% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collocate.R
\name{collocate}
\alias{collocate}
\title{A collocation function}
\usage{
collocate(corpus, keyword, collocation_width = 4, tidy = TRUE)
}
\arguments{
\item{corpus}{A tm corpus}

\item{keyword}{The keyword that you wish to find collocates for}

\item{collocation_width}{The number of neighbouring words on each side}

\item{tidy}{Tidy the corpus - remove punctuation, numbers, empty lines (optional) and whitespace, convert to lowercase}

\item{remove_empty_lines}{Remove empty lines from the corpus (default: TRUE)}
}
\description{
Takes a tm corpus as input, returns a list of matrices containing collocates as output.
}
\examples{
library(tm)
data("crude")
collocate(crude,"oil", collocation_width = 6, tidy = TRUE)
}
\keyword{collocates}
\keyword{concordance}

