% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collocateR.R
\docType{package}
\name{collocateR}
\alias{collocateR}
\alias{collocateR-package}
\title{collocateR: a package that calculates significance measures for collocates}
\description{
This package acts as an effective plugin for the Quanteda and Tidytext packages 
It applies statistical measures to collocates, generated as keywords in context
}
\examples{
# Example usage:
library(collocateR)
}
