% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_collocates_corpus.R
\name{collocates_corpus}
\alias{collocates_corpus}
\title{returns a quanteda corpus of collocates from a quanteda kwic file}
\usage{
collocates_corpus(kwic)
}
\arguments{
\item{vector}{A vector of text files}
}
\description{
returns a quanteda corpus of collocates from a quanteda kwic file
}
\keyword{tidy}

