% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Compare_Kwics.R
\name{compare_kwics}
\alias{compare_kwics}
\title{Compare two sets of keyword in context matrices}
\usage{
compare_kwics(original_corpus, kwic_list, item_selection1, item_selection2,
  weighting = "raw", remove_stops = FALSE, tidy = FALSE)
}
\arguments{
\item{kwic_list}{A list of kwic matrices produced using the 'kwic' function in collocateR}

\item{item_selection1}{First subset to be compared}

\item{item_selection2}{Second subset to be compared}

\item{weighting}{For sorting. 'raw' = raw frequency; 'normalised' = Normalised frequency (within the KWIC selection; 'tfidf' = Term Frequency-Inverse Docmuent Frequency}

\item{remove_stops}{Remove stopwords}

\item{tidy}{Tidy the corpus}

\item{corpus}{A tm corpus (tidied)}
}
\description{
Takes a 'kwic' matrix, plus two nominated subsetting vecors, returns a list of 2 scores for significant words, sorted by frequency or through a 'term frequency-inverse document frequency' ('tf-idf') that compares normalised tf in the kwics to idf for the corpus as a whole.
}
\examples{
library(tm)
data("crude")
oilkwic <- kwic(crude, "oil", collocation_width = 6, tidy = TRUE)
compare_kwics(crude, oilkwic, 1:10, 11:20, weighting = "normal", remove_stops = FALSE, tidy = FALSE)

}
\keyword{collocates}
\keyword{concordance}

