% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corpus_find.R
\name{corpus_find}
\alias{corpus_find}
\title{A small function using tm to find a keyword in a corpus}
\usage{
corpus_find(corpus, keyword)
}
\arguments{
\item{corpus}{A corpus}

\item{keyword}{A search term}
}
\description{
Returns a new corpus composed of all the documents in the original corpus that contain a keyword
}
\keyword{search}
\keyword{tm}

