% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_collocates.R
\name{count_collocates}
\alias{count_collocates}
\title{An internal function for counting collocates. Accepts a collsDB and returns a tibble containing collocate counts}
\usage{
count_collocates(collsDB)
}
\arguments{
\item{collsDB}{A collocates list, combined with relevant counts, generated by save_collocates}
}
\description{
An internal function for counting collocates. Accepts a collsDB and returns a tibble containing collocate counts
}
