% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_multigrams.R
\name{count_multigrams}
\alias{count_multigrams}
\title{An internal function for counting multigrams. Accepts a collsDB and returns a tibble containing collocate counts}
\usage{
count_multigrams(collsDB, ngram)
}
\arguments{
\item{collsDB}{A collocates list, combined with relevant counts, generated by save_collocates}
}
\description{
An internal function for counting multigrams. Accepts a collsDB and returns a tibble containing collocate counts
}
