% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_unigrams.R
\name{count_unigrams}
\alias{count_unigrams}
\title{An internal function for counting collocates. Accepts a collsDB and returns a tibble containing collocate counts}
\usage{
count_unigrams(collsDB)
}
\arguments{
\item{collsDB}{A collocates list, combined with relevant counts, generated by save_collocates}
}
\description{
An internal function for counting collocates. Accepts a collsDB and returns a tibble containing collocate counts
}
