% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collocate-early.R
\name{early_collocates}
\alias{early_collocates}
\title{A function that tackles early words}
\usage{
early_collocates(collocation_width, early, input_vector, keyword,
  keyword_locations)
}
\arguments{
\item{collocation_width}{The collocation width}

\item{early}{The early word's location}

\item{input_vector}{The vector that is being processed}

\item{keyword}{The relevant keyword}

\item{keyword_locations}{Keyword locations within the doc}
}
\description{
Addresses words that are less than a collocation_width away from the document's beginning
}
\keyword{collocates}
\keyword{concordance}

