% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualise_dispersion_plot_grid.R
\name{faceted_dispersion_plot}
\alias{faceted_dispersion_plot}
\title{Take a dataframe of word locations and create a faceted dispersion plot}
\usage{
faceted_dispersion_plot(vectors, keyword, panel_labels = names(vectors),
  plottitle = NULL)
}
\arguments{
\item{vectors}{Vectors of text files}

\item{keyword}{A keyword}

\item{panel_labels}{The labels for each vector}

\item{plottitle}{A title for the plot}
}
\description{
Take a dataframe of word locations and create a faceted dispersion plot
}
\examples{
test <- faceted_dispersion_plot(csr2, "risk", years, plottitle = "Word locations for 'risk' in BP's CSR Reports 1998-2014")
}
\keyword{dispersion}
\keyword{ggplot}
\keyword{plot}

