% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_collocates.R
\name{find_collocates}
\alias{find_collocates}
\title{returns a list containing node locations, window locations, the node and its hash, and a doc_table}
\usage{
find_collocates(document, window, node, remove_stops = TRUE)
}
\arguments{
\item{document}{The document to be analysed}

\item{window}{The size of the collocate window on each side of tne node}

\item{node}{A key word or phrase}

\item{remove_stops}{If TRUE, stopwords are removed (stopwords derived from tidytext package)}
}
\description{
returns a list containing node locations, window locations, the node and its hash, and a doc_table
}
\keyword{collocates}
\keyword{kwic}
