% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_freqs.R
\name{get_freqs}
\alias{get_freqs}
\title{Extract frequencies for the keywords in context and the full document}
\usage{
get_freqs(doc, keyword, window = 6, ngram = 1,
  remove_stopwords = TRUE, span = "both")
}
\arguments{
\item{doc}{A character vector or list of character vectors}

\item{keyword}{A key word or phrase to test}

\item{window}{The number of context words to be displayed around the keyword Default 6}

\item{ngram}{The size of phrases the frequencies of which we are to test (so, unigram = 1, bigram = 2, trigram = 3 etc)}

\item{remove_stopwords}{Remove stopwords, derived from Quanteda's list}

\item{span}{Whether to include a window's width of words to the left of the keyword, to the right or on both sides}
}
\description{
Extract frequencies for the keywords in context and the full document
}
\keyword{collocates,}
\keyword{information,}
\keyword{kwic}
\keyword{mutual}
