% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mi3.R
\name{get_mi3}
\alias{get_mi3}
\title{mutual information cubed drawing on information in Church and Hanks 1989 http://www.aclweb.org/anthology/J90-1003}
\usage{
get_mi3(doc, keyword, window = 6, ngram = 1, remove_stopwords = TRUE,
  min_count = 2)
}
\arguments{
\item{doc}{Acharacter vector or list of character vectors}

\item{keyword}{A key word or phrase to test}

\item{window}{The number of context words to be displayed around the keyword Default 6}

\item{ngram}{The size of phrases the frequencies of which we are to test (so, unigram = 1, bigram = 2, trigram = 3 etc)}

\item{remove_stopwords}{Remove stopwords, derived from Quanteda's list}

\item{min_count}{Collocates that occur fewer times than floor will be removed}
}
\description{
mutual information cubed drawing on information in Church and Hanks 1989 http://www.aclweb.org/anthology/J90-1003
}
\keyword{collocates,}
\keyword{information,}
\keyword{kwic}
\keyword{mutual}
