% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_npmi.R
\name{get_npmi}
\alias{get_npmi}
\title{normalised pointwise mutual information: a significance test for keywords in context, normalised between 1 and -1}
\usage{
get_npmi(document, pattern, window = 6, ngram = 1, floor = 3,
  remove_stopwords = TRUE, remove_numerals = TRUE, cache = FALSE)
}
\arguments{
\item{document}{Acharacter vector or list of character vectors}

\item{window}{The number of context words to be displayed around the keyword Default 5}

\item{ngram}{The size of phrases the frequencies of which we are to test (so, unigram = 1, bigram = 2, trigram = 3 etc)}

\item{floor}{Collocates that occur fewer times than floor will be removed}

\item{remove_stopwords}{Remove stopwords from the document (based on tidytext's stopwords data). Default TRUE.}

\item{remove_numerals}{Remove numerals}

\item{cache}{Organising collocates is the most time-consuming step in calculating frequencies and other collocation algorithms. The memoise package is used to cache specific iterations of this process. Default FALSE.}

\item{node}{A key word or phrase to test}
}
\description{
normalised pointwise mutual information: a significance test for keywords in context, normalised between 1 and -1
}
\keyword{collocates,}
\keyword{information,}
\keyword{kwic}
\keyword{mutual}
