% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pmi.R
\name{get_pmi}
\alias{get_pmi}
\title{pointwise mutual information: internal function (for use in pmi and npmi functions)}
\usage{
get_pmi(document, freqs)
}
\arguments{
\item{document}{A collDB, produced through save_collocates, or a text file}

\item{freqs}{collocate and document frequencies}
}
\description{
pointwise mutual information: internal function (for use in pmi and npmi functions)
}
\keyword{collocates,}
\keyword{information,}
\keyword{kwic}
\keyword{mutual}
