% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_get_freqs.R
\name{internal_get_freqs}
\alias{internal_get_freqs}
\title{Extract frequencies for the keywords in context and the full document}
\usage{
internal_get_freqs(doc, keyword, window = window, ngram = ngram,
  min_count = min_count, cache = cache)
}
\arguments{
\item{doc}{A character vector or list of character vectors}

\item{keyword}{A key word or phrase to test}

\item{window}{The number of context words to be displayed around the keyword Default 6}

\item{ngram}{The size of phrases the frequencies of which we are to test (so, unigram = 1, bigram = 2, trigram = 3 etc)}

\item{min_count}{Collocates that occur fewer times than floor will be removed}

\item{cache}{Getting frequencies is the most time-consuming step in calculating frequencies and other collocation algorithms. The memoise package is used to cache specific iterations of this process. Default FALSE.}
}
\description{
Extract frequencies for the keywords in context and the full document
}
\keyword{collocates,}
\keyword{information,}
\keyword{kwic}
\keyword{mutual}
