% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_manage_keyword.R
\name{internal_manage_keyword}
\alias{internal_manage_keyword}
\title{pointwise mutual information: internal function (for use in pmi and npmi functions)}
\usage{
internal_manage_keyword(doc, keyword)
}
\arguments{
\item{doc}{Acharacter vector or list of character vectors}

\item{keyword}{A key word or phrase to test}
}
\description{
pointwise mutual information: internal function (for use in pmi and npmi functions)
}
\keyword{collocates,}
\keyword{information,}
\keyword{kwic}
\keyword{mutual}
