% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iscollDB.R
\name{is.collDB}
\alias{is.collDB}
\title{internal function to test whether an object is a collDB}
\usage{
is.collDB(document)
}
\arguments{
\item{document}{The document to be analysed}
}
\description{
internal function to test whether an object is a collDB
}
