% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collocate.R
\name{kwic}
\alias{kwic}
\title{A keywords in context function}
\usage{
kwic(corpus, keyword, collocation_width = 4, tidy = FALSE, exact = TRUE)
}
\arguments{
\item{corpus}{A tm corpus}

\item{keyword}{The keyword that you wish to find collocates for}

\item{collocation_width}{The number of neighbouring words on each side}

\item{tidy}{Tidy the corpus - remove punctuation, numbers, empty lines and whitespace, convert to lowercase}

\item{exact}{FALSE to use 'grep' to find all words containg your keyword phrase or TRUE to use 'which' to find the exact keyword (default: TRUE (is much faster))}
}
\description{
Takes a tm corpus as input, returns a list of matrices containing Keywords in Context as output.
}
\examples{
library(tm)
data("crude")
kwic(crude,"oil", collocation_width = 6, tidy = TRUE)
}
\keyword{collocates}
\keyword{concordance}

