% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_kwic_scores.R
\name{kwic_scores}
\alias{kwic_scores}
\title{returns a wordscore dataframe of collocates from a quanteda corpus file}
\usage{
kwic_scores(corpus, keywords, window = 5)
}
\arguments{
\item{window}{The width of the kwic vector}

\item{vector}{A vector of text files}

\item{keyword}{A keyword to feed to the kwic vector}
}
\description{
returns a wordscore dataframe of collocates from a quanteda corpus file
}
\keyword{tidy}

