% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_regenerate_corpus.R
\name{kwic_to_vect}
\alias{kwic_to_vect}
\title{returns a vector from a kwic file}
\usage{
kwic_to_vect(corpus, keywords, window)
}
\arguments{
\item{window}{The width of the kwic vector}

\item{vector}{A vector of text files}

\item{keyword}{A keyword to feed to the kwic vector}
}
\description{
returns a vector from a kwic file
}
\keyword{corpus}
\keyword{kwic}

