% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collocate-late.R
\name{late_collocates}
\alias{late_collocates}
\title{A function that tackles late words}
\usage{
late_collocates(collocation_width, doc_length, input_vector, keyword,
  keyword_locations, late)
}
\arguments{
\item{collocation_width}{The collocation width}

\item{doc_length}{The document's word count}

\item{input_vector}{The vector that is being processed}

\item{keyword}{The relevant keyword}

\item{keyword_locations}{Keyword locations within the doc}

\item{late}{The late word's location}
}
\description{
Addresses words that are less than a collocation_width away from the document's end
}
\keyword{collocates}
\keyword{concordance}

