% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/background_make_dfm.R
\name{make_dfm}
\alias{make_dfm}
\title{A generic function for making a dfm}
\usage{
make_dfm(input, ngram)
}
\arguments{
\item{input}{A vector or a corpus}

\item{ngram}{The ngram types}
}
\description{
A generic function for making a dfm
}
\keyword{corpus}
\keyword{information}
\keyword{kwic}
\keyword{mutual}

