% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mi.R
\name{mi}
\alias{mi}
\title{mutual information drawing on information in Church and Hanks 1989 http://www.aclweb.org/anthology/J90-1003}
\usage{
mi(doc, keyword, window = 6, ngram = 1, remove_stopwords = TRUE,
  min_count = 2, span = "both")
}
\arguments{
\item{doc}{Acharacter vector or list of character vectors}

\item{keyword}{A key word or phrase to test}

\item{window}{The number of context words to be displayed around the keyword Default 6}

\item{ngram}{The size of phrases the frequencies of which we are to test (so, unigram = 1, bigram = 2, trigram = 3 etc)}

\item{remove_stopwords}{Remove stopwords, derived from Quanteda's list}

\item{min_count}{Collocates that occur fewer times than floor will be removed}

\item{span}{Whether to include a window's width of words to the left of the keyword, to the right or on both sides}
}
\description{
mutual information drawing on information in Church and Hanks 1989 http://www.aclweb.org/anthology/J90-1003
}
\keyword{collocates,}
\keyword{information,}
\keyword{kwic}
\keyword{mutual}
