% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mins_to_maxs.R
\name{mins_to_maxs}
\alias{mins_to_maxs}
\title{An internal function for returning a list of locations, including notes and running all contiguous words together}
\usage{
mins_to_maxs(collDBdoc)
}
\arguments{
\item{collsDB}{A collocates list, combined with relevant counts, generated by save_collocates}
}
\description{
An internal function for returning a list of locations, including notes and running all contiguous words together
}
