% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collocate-middle.R
\name{normal_collocates}
\alias{normal_collocates}
\title{A function that isolates keywords and neighbours}
\usage{
normal_collocates(collocation_width, input_vector, keyword_locations, keyword)
}
\arguments{
\item{collocation_width}{The collocation width}

\item{input_vector}{The vector that is being processed}

\item{keyword_locations}{Keyword locations within the doc}

\item{keyword}{The relevant keyword}
}
\description{
Addresses words that are less than a collocation_width away from the document's end
}
\keyword{collocates}
\keyword{concordance}

