% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_npmi.R
\name{npmi}
\alias{npmi}
\title{Generates ngrams and delivers pmi scores by working from kwic file back to the full corpus (more efficient than generating large ngrams across the whole corpus)}
\usage{
npmi(vect, keyword, ngram = 1:2, window = 5, cutoff = 3, stem = FALSE,
  cached_tokens = NULL)
}
\arguments{
\item{vect}{A vector of text files or a corpus}

\item{keyword}{A keyword to feed to the kwic vector}

\item{ngram}{The ngram types}

\item{window}{The width of the kwic vector}

\item{cutoff}{Only collocation frequencies above the cutoff will be included}

\item{cached_tokens}{Deliver a tokenized vector for the whole corpus (good for repeating analyses)}
}
\description{
Generates ngrams and delivers pmi scores by working from kwic file back to the full corpus (more efficient than generating large ngrams across the whole corpus)
}
\keyword{corpus}
\keyword{information}
\keyword{kwic}
\keyword{mutual}

