% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npmi_for_multigrams.R
\name{npmi_multigrams}
\alias{npmi_multigrams}
\title{Generates ngrams and delivers pmi scores by working from kwic file back to the full corpus (more efficient than generating large ngrams across the whole corpus)}
\usage{
npmi_multigrams(vect, keyword, ngram = 1:2, window = 5, cutoff = 3)
}
\arguments{
\item{vect}{A vector of text files or a corpus}

\item{keyword}{A keyword to feed to the kwic vector}

\item{ngram}{The ngram types}

\item{window}{The width of the kwic vector}
}
\description{
Generates ngrams and delivers pmi scores by working from kwic file back to the full corpus (more efficient than generating large ngrams across the whole corpus)
}
\keyword{corpus}
\keyword{information}
\keyword{kwic}
\keyword{mutual}

