% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/significance_observed_expected.R
\name{observed_expected}
\alias{observed_expected}
\title{A observed/expected word frequency function}
\usage{
observed_expected(origcorpus, collocates, tidy = TRUE, threshold = 3)
}
\arguments{
\item{origcorpus}{A collocation matrix}

\item{tidy}{Tidy the corpus - remove punctuation, numbers, empty lines (optional) and whitespace, convert to lowercase}

\item{threshold}{Remove low frequency words (default = 3)}
}
\description{
Takes a tm corpus and a KWIC matrix as inputs, returns a list of matrices containing observed/expected ratios for each word in the KWIC matrix
}
\examples{
library(tm)
data("crude")
oil_kwic <- collocate(crude,"oil", collocation_width = 6, tidy = TRUE)
observed_expected(crude, oil_kwic)
}
\keyword{collocates}
\keyword{concordance}
\keyword{observed/expected}

