% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Process_MIM_Ngram_Scores.R
\name{pmi_score}
\alias{pmi_score}
\title{Calculates a pointwise mutual information score for ngrams in a corpus: log2(prob(xy)/prob(x)prob(y))}
\usage{
pmi_score(vect, keyword, window = 5, ngram = 1, cutoff = 3)
}
\arguments{
\item{vect}{A vector of text files or a corpus}

\item{keyword}{A keyword to feed to the kwic vector}

\item{window}{The width of the kwic vector}

\item{ngram}{The ngram types}

\item{cutoff}{A floor for infrequent terms}
}
\description{
Calculates a pointwise mutual information score for ngrams in a corpus: log2(prob(xy)/prob(x)prob(y))
}
\keyword{corpus}
\keyword{information}
\keyword{kwic}
\keyword{mutual}

