% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_read_texts.R
\name{read_texts}
\alias{read_texts}
\title{Read a Quanteda corpus into R}
\usage{
read_texts(directory, tidy = TRUE, remove_stopwords = TRUE,
  as_corpus = FALSE)
}
\arguments{
\item{directory}{A directory}

\item{tidy}{Tidy the corpus - remove all punctuation, numbers, empty lines and whitespace, convert to lowercase}

\item{remove_stopwords}{Remove all stopwords. For a list of stopwords input quanteda::stopwords("english")}

\item{as_corpus}{return as a quanteda corpus}
}
\description{
Read text files into R and tidy them.
}
\examples{
read_texts("~/data/", tidy=TRUE)
}
\keyword{import}
\keyword{texts}

