% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/background_remove_stopwords.R
\name{remove_stops}
\alias{remove_stops}
\title{Removes stopwords from a vector of text files, using quanteda's stopwords}
\usage{
remove_stops(vectors)
}
\arguments{
\item{vectors}{A vector of text files}
}
\description{
Removes stopwords from a vector of text files, using quanteda's stopwords
}
\keyword{tidy}

