% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sanitise_doc.R
\name{sanitise_doc}
\alias{sanitise_doc}
\title{Remove certain puntuations (apostrophes etc) and then retain only letters and spaces}
\usage{
sanitise_doc(doc)
}
\arguments{
\item{doc}{A character vector}
}
\description{
Remove certain puntuations (apostrophes etc) and then retain only letters and spaces
}
\keyword{sanitise}
\keyword{tidy}
