% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_collocates.R
\name{save_collocates}
\alias{save_collocates}
\title{returns a list containing node locations, window locations, the node and its hash, and a doc_table}
\usage{
save_collocates(document, window, node, remove_stops = TRUE,
  remove_numerals = TRUE, remove_punct = TRUE)
}
\arguments{
\item{document}{The document to be analysed}

\item{window}{The size of the collocate window on each side of tne node}

\item{node}{A key word or phrase}

\item{remove_stops}{If TRUE, stopwords are removed (stopwords derived from quanteda package)}

\item{remove_numerals}{If TRUE, numerals are removed}

\item{remove_punct}{If TRUE, puntuation is removed}
}
\description{
returns a list containing node locations, window locations, the node and its hash, and a doc_table
}
\keyword{collocates}
\keyword{kwic}
