% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_kwic.R
\name{simple_kwic}
\alias{simple_kwic}
\title{This function does not aim to replace the more thorough function from the Quanteda package. Instead it simply returns a tibble of KWICs arranged by row.
Internal formula for processing multigrams}
\usage{
simple_kwic(document, pattern, window = 5, ngram = 1,
  remove_stopwords = TRUE, remove_numerals = TRUE, cache = FALSE)
}
\arguments{
\item{document}{A character vector or list of character vectors}

\item{pattern}{A character vector containing a keyword}

\item{window}{The number of context words to be displayed around the keyword Default 5}

\item{ngram}{The size of phrases the frequencies of which we are to test (so, unigram = 1, bigram = 2, trigram = 3 etc)}

\item{remove_stopwords}{Remove stopwords from the document (drawn from tidytext's stopwords data). Default TRUE.}

\item{remove_numerals}{Remove numerals}

\item{cache}{Organising collocates is the most time-consuming step in calculating frequencies and other collocation algorithms. The memoise package is used to cache specific iterations of this process. Deault FALSE.}
}
\description{
This function does not aim to replace the more thorough function from the Quanteda package. Instead it simply returns a tibble of KWICs arranged by row.
Internal formula for processing multigrams
}
\keyword{context}
\keyword{in}
\keyword{keywords}
