% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_kwics.R
\name{simple_kwics}
\alias{simple_kwics}
\title{Calculates a pointwise mutual information score for ngrams in a corpus: log2(prob(xy)/prob(x)prob(y))}
\usage{
simple_kwics(vect, keyword, window)
}
\arguments{
\item{vect}{A vector of text files}

\item{keyword}{A keyword to feed to the kwic vector}

\item{window}{The width of the kwic vector}
}
\description{
Calculates a pointwise mutual information score for ngrams in a corpus: log2(prob(xy)/prob(x)prob(y))
}
\keyword{corpus}
\keyword{information}
\keyword{kwic}
\keyword{mutual}

