% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R, R/stop_words_importer.R
\name{stop_words}
\alias{stop_words}
\alias{stop_words}
\title{Various lexicons for English stop words}
\format{A data frame with 1149 rows and 2 variables:
\describe{
 \item{word}{An English word}
 \item{lexicon}{The source of the stop word. Either "onix", "SMART", or "snowball"}
 }}
\source{
\itemize{
\item \url{http://www.lextek.com/manuals/onix/stopwords1.html}
\item \url{http://jmlr.csail.mit.edu/papers/volume5/lewis04a/a11-smart-stop-list/english.stop}
\item \url{http://snowball.tartarus.org/algorithms/english/stop.txt}
}

\url{https://github.com/juliasilge/tidytext/tree/master/data-raw}
}
\usage{
stop_words()

stop_words()
}
\description{
English stop words from three lexicons, as a data frame.
The onix and SMART sets are pulled from the tm package. Note
that words with non-ASCII characters have been removed.

A stopwords script from tidytext (https://github.com/juliasilge/tidytext/blob/3006065b79ff4cb3f2c4edc37914176448105943/data-raw/setup_data.R)
}
