% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/background-tidy.R
\name{tidy}
\alias{tidy}
\title{Tidies a vector of text files}
\usage{
tidy(vector)
}
\arguments{
\item{vector}{A vector of text files}
}
\description{
Tidies a vector of text files
}
\keyword{tidy}

