% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_word_locations.R
\name{word_locations}
\alias{word_locations}
\title{Turn a text vector into a data frame of keyword locations}
\usage{
word_locations(vect, keyword, vectorname)
}
\arguments{
\item{vect}{A text vector}

\item{keyword}{A keyword}

\item{A}{vectorname The Vector name or an assigned vectorname}
}
\description{
Turn a text vector into a data frame with three columns: 1. the filename or assigned name; 2. the word location; 3. keyword match (0 for does not match, 1 for match)
}
\keyword{locations}
\keyword{word}

