% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_z_score.R
\name{z_score}
\alias{z_score}
\title{A z-score word frequency function, operationalising calculations contained in Barnbrook Geoff and others, Collocation: Applications and Implications (Palgrave Macmillan 2013), Chapter 3.}
\usage{
z_score(origcorpus, collocates, tidy = TRUE, threshold = 3)
}
\arguments{
\item{origcorpus}{A collocation matrix}

\item{tidy}{Tidy the corpus - remove punctuation, numbers, empty lines (optional) and whitespace, convert to lowercase}

\item{threshold}{Remove low frequency words (default = 3)}
}
\description{
Takes a tm corpus and a KWIC matrix as inputs, returns a list of matrices containing observed/expected ratios for each word in the KWIC matrix
}
\examples{
library(tm)
data("crude")
oil_kwic <- kwic(crude,"oil", collocation_width = 6, tidy = TRUE)
z_score(crude, oil_kwic)
}
\keyword{collocates}
\keyword{concordance}
\keyword{z-score}

