% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zscore.R
\name{zscore}
\alias{zscore}
\title{A z-score calculates a probability score that compares the observed collocate frequency to its expected value}
\usage{
zscore(doc, keyword, window = 6, ngram = 1, remove_stopwords = TRUE,
  min_count = 2, span = "both")
}
\arguments{
\item{doc}{A character vector or list of character vectors}

\item{keyword}{A character vector containing a keyword}

\item{window}{The number of context words to be displayed around the keyword Default 6}

\item{ngram}{The size of phrases the frequencies of which we are to test (so, unigram = 1, bigram = 2, trigram = 3 etc)}

\item{remove_stopwords}{Remove stopwords, derived from Quanteda's list}

\item{min_count}{The minimum frequency to include in the score}

\item{span}{Whether to include a window's width of words to the left of the keyword, to the right or on both sides}
}
\description{
A z-score calculates a probability score that compares the observed collocate frequency to its expected value
}
\keyword{collocates,}
\keyword{kwic}
\keyword{z_score,}
