% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_freqs.R
\name{get_freqs}
\alias{get_freqs}
\title{Get raw frequencies for ngrams within the document}
\usage{
get_freqs(document, ngrams = 1)
}
\arguments{
\item{document}{A collDB, produced through save_collocates, or a text file}

\item{ngrams}{The length of the ngrams to be tested}
}
\description{
Get raw frequencies for ngrams within the document
}
\keyword{frequencies}
