% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_freqs.R
\name{get_freqs}
\alias{get_freqs}
\title{Get raw frequencies for ngrams within the document}
\usage{
get_freqs(document, pattern, window = 6, ngram = 1,
  remove_stopwords = TRUE, cache = FALSE)
}
\arguments{
\item{pattern}{A character vector containing a keyword}

\item{window}{The number of context words to be displayed around the keyword Default 5}

\item{ngram}{The size of phrases the frequencies of which we are to test (so, unigram = 1, bigram = 2, trigram = 3 etc)}

\item{remove_stopwords}{Remove stopwords from the document (based on tidytext's stopwords data). Default TRUE.}

\item{cache}{Organising collocates is the most time-consuming step in calculating frequencies and other collocation algorithms. The memoise package is used to cache specific iterations of this process. Deault FALSE.}

\item{doc}{A character vector or list of character vectors}
}
\description{
Get raw frequencies for ngrams within the document
}
\keyword{frequencies}
