% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_npmi.R
\name{get_npmi}
\alias{get_npmi}
\title{pointwise mutual information: internal function (for use in pmi and npmi functions)}
\usage{
get_npmi(doc, keyword, window = 6, ngram = 1, min_count = 2,
  cache = FALSE)
}
\arguments{
\item{doc}{Acharacter vector or list of character vectors}

\item{keyword}{A key word or phrase to test}

\item{window}{The number of context words to be displayed around the keyword Default 6}

\item{ngram}{The size of phrases the frequencies of which we are to test (so, unigram = 1, bigram = 2, trigram = 3 etc)}

\item{min_count}{Collocates that occur fewer times than floor will be removed}

\item{cache}{Organising collocates is the most time-consuming step in calculating frequencies and other collocation algorithms. The memoise package is used to cache specific iterations of this process. Default FALSE.}
}
\description{
pointwise mutual information: internal function (for use in pmi and npmi functions)
}
\keyword{collocates,}
\keyword{information,}
\keyword{kwic}
\keyword{mutual}
