% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_get_pmi.R
\name{internal_get_pmi}
\alias{internal_get_pmi}
\title{pointwise mutual information: internal function (for use in pmi and npmi functions)}
\usage{
internal_get_pmi(doc, keyword, window, ngram, remove_stopwords, min_count)
}
\arguments{
\item{doc}{A character vector or list of character vectors}

\item{keyword}{A key word or phrase to test}

\item{window}{The number of context words to be displayed around the keyword Default 5}

\item{ngram}{The size of phrases the frequencies of which we are to test (so, unigram = 1, bigram = 2, trigram = 3 etc)}

\item{remove_stopwords}{Remove stopwords, derived from Quanteda's list}

\item{min_count}{Collocates that occur fewer times than floor will be removed}
}
\description{
pointwise mutual information: internal function (for use in pmi and npmi functions)
}
\keyword{collocates,}
\keyword{information,}
\keyword{kwic}
\keyword{mutual}
