% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npmi.R
\name{npmi}
\alias{npmi}
\title{normalised pointwise mutual information: a significance test for keywords in context, normalised between 1 and -1}
\usage{
npmi(document, floor = 3, ngrams = 1, window, node, remove_stops = TRUE,
  remove_numerals = TRUE, remove_punct = TRUE)
}
\arguments{
\item{document}{A collDB, produced through save_collocates, or a text file}

\item{floor}{Collocates that occur fewer times than floor will be removed}

\item{window}{OPTIONAL The size of the collocate window on each side of tne node. Not required if importing a collDB from save_collocates}

\item{node}{OPTIONAL A key word or phrase. Not required if importing a collDB from save_collocates}

\item{remove_stops}{OPTIONAL If TRUE, stopwords are removed (stopwords derived from tidytext package). Not required if importing a collDB from save_collocates}

\item{remove_numerals}{OPTIONAL If TRUE, numerals are removed. Not required if importing a collDB from save_collocates}

\item{remove_punct}{OPTIONAL If TRUE, puntuation is removed. Not required if importing a collDB from save_collocates}
}
\description{
normalised pointwise mutual information: a significance test for keywords in context, normalised between 1 and -1
}
\keyword{collocates,}
\keyword{information,}
\keyword{kwic}
\keyword{mutual}
