% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmi.R
\name{pmi}
\alias{pmi}
\title{pointwise mutual information: a significance test for keywords in context}
\usage{
pmi(document, floor = 3, ngrams = 1)
}
\arguments{
\item{document}{A collDB, produced through save_collocates, or a text file}

\item{floor}{Collocates that occur fewer times than floor will be removed}

\item{ngrams}{The length of the ngrams to be tested}

\item{window}{OPTIONAL The size of the collocate window on each side of tne node. Not required if importing a collDB from save_collocates}

\item{node}{OPTIONAL A key word or phrase. Not required if importing a collDB from save_collocates}

\item{remove_stops}{OPTIONAL If TRUE, stopwords are removed (stopwords derived from tidytext package). Not required if importing a collDB from save_collocates}

\item{remove_numerals}{OPTIONAL If TRUE, numerals are removed. Not required if importing a collDB from save_collocates}

\item{remove_punct}{OPTIONAL If TRUE, puntuation is removed. Not required if importing a collDB from save_collocates}
}
\description{
pointwise mutual information: a significance test for keywords in context
}
\keyword{collocates,}
\keyword{information,}
\keyword{kwic}
\keyword{mutual}
