% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_duplicates.R
\name{remove_duplicates}
\alias{remove_duplicates}
\title{Extract frequencies for the keywords in context and the full document}
\usage{
remove_duplicates(collocates, keyword = keyword, window = window)
}
\arguments{
\item{collocates}{A kwic matrix}
}
\description{
Extract frequencies for the keywords in context and the full document
}
\keyword{collocates,}
\keyword{information,}
\keyword{kwic}
\keyword{mutual}
