% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corpus-tidy.R
\name{tidy}
\alias{tidy}
\title{Uses tm to tidy a corpus}
\usage{
tidy(corpus_raw)
}
\arguments{
\item{corpus_raw}{A corpus that has not been processed}

\item{remove_empty_lines}{Remove empty lines from the corpus. Default TRUE}
}
\description{
Addresses words that are less than a collocation_width away from the document's end
}
\keyword{collocates}
\keyword{concordance}

