% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_score.R
\name{z_score}
\alias{z_score}
\title{A z-score calculates a probability score that compares the observed collocate frequency to its expected value}
\usage{
z_score(document, pattern, window = 5, ngram = 1, floor = 3,
  remove_stopwords = TRUE, cache = FALSE)
}
\arguments{
\item{document}{A character vector or list of character vectors}

\item{pattern}{A character vector containing a keyword}

\item{window}{The number of context words to be displayed around the keyword Default 5}

\item{ngram}{The size of phrases the frequencies of which we are to test (so, unigram = 1, bigram = 2, trigram = 3 etc)}

\item{remove_stopwords}{Remove stopwords from the document (based on tidytext's stopwords data). Default TRUE.}

\item{cache}{Organising collocates is the most time-consuming step in calculating frequencies and other collocation algorithms. The memoise package is used to cache specific iterations of this process. Default FALSE.}
}
\description{
A z-score calculates a probability score that compares the observed collocate frequency to its expected value
}
\keyword{collocates,}
\keyword{kwic}
\keyword{z_score,}
